Extra tests and refactors
