+++ saleor/graphql/order/tests/test_order.py
@@ -3382,10 +3382,14 @@ def test_order_update_shipping_incorrect_shipping_method(
 
 
 def test_order_update_shipping_excluded_shipping_method_zip_code(
-    staff_api_client, permission_manage_orders, order, staff_user, shipping_method
+    staff_api_client,
+    permission_manage_orders,
+    order,
+    staff_user,
+    shipping_method_excldued_by_zip_code,
 ):
-    order.shipping_method = shipping_method
-    shipping_total = shipping_method.channel_listings.get(
+    order.shipping_method = shipping_method_excldued_by_zip_code
+    shipping_total = shipping_method_excldued_by_zip_code.channel_listings.get(
         channel_id=order.channel_id,
     ).get_total()
 
@@ -3396,7 +3400,10 @@ def test_order_update_shipping_excluded_shipping_method_zip_code(
 
     query = ORDER_UPDATE_SHIPPING_QUERY
     order_id = graphene.Node.to_global_id("Order", order.id)
-    variables = {"order": order_id, "shippingMethod": None}
+    method_id = graphene.Node.to_global_id(
+        "ShippingMethod", shipping_method_excldued_by_zip_code.id
+    )
+    variables = {"order": order_id, "shippingMethod": method_id}
     response = staff_api_client.post_graphql(
         query, variables, permissions=[permission_manage_orders]
     )
