+++ saleor/order/utils.py
@@ -331,20 +331,6 @@ def get_prices_of_discounted_specific_product(
     return line_prices
 
 
-def check_zip_code_in_excluded_range(code, start, end):
-    # TODO: do actual logic on checking if the zip code is in range
-    return True
-
-
-def check_shipping_method_for_zip_code(customer_zip_code, method: ShippingMethod):
-    for zip_code in method.zip_code_rules.all():
-        if check_zip_code_in_excluded_range(
-            customer_zip_code, zip_code.start, zip_code.end
-        ):
-            return True
-    return False
-
-
 def get_products_voucher_discount_for_order(order: Order) -> Money:
     """Calculate products discount value for a voucher, depending on its type."""
     prices = None
